test: api test partial
