+++ trss/api_test.go
@@ -3,261 +3,119 @@ package transmissionrss_test
 import (
 	"encoding/json"
 	"errors"
-	"fmt"
+
+	// "fmt"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
 	"os"
-	"testing"
 
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+
+	// helper "github.com/iben12/transmission-rss-go/tests/helper"
 	"github.com/iben12/transmission-rss-go/tests/mocks"
 	trss "github.com/iben12/transmission-rss-go/trss"
-	"github.com/rs/zerolog"
-	"github.com/stretchr/testify/assert"
-	"gorm.io/gorm"
 )
 
-func TestApi(t *testing.T) {
-	assert := assert.New(t)
-	rssAddress := "https://example.rss/feed"
-	os.Setenv("RSS_FEED_ADDRESS", rssAddress)
-	mockFeeds := &mocks.MockFeeds{}
-	trss.FeedService = mockFeeds
-	mockEpisodes := &mocks.MockEpisodes{}
-	trss.EpisodeService = mockEpisodes
-	mockTrs := &mocks.MockTransmissionService{}
-	trss.TrsService = mockTrs
-	trss.Logger = zerolog.New(ioutil.Discard).With().Timestamp().Logger()
+type ApiEndpoint func(w http.ResponseWriter, r *http.Request)
 
-	t.Run("Feeds endpoint", func(t *testing.T) {
-		req := httptest.NewRequest(http.MethodGet, "/", nil)
-		w := httptest.NewRecorder()
-		api := new(trss.Api)
+func sendApiRequest(endpoint ApiEndpoint, method string, url string) (*http.Response, []byte) {
+	req := httptest.NewRequest(method, url, nil)
+	w := httptest.NewRecorder()
 
-		api.Feeds(w, req)
+	endpoint(w, req)
 
-		res := w.Result()
+	res := w.Result()
 
-		defer res.Body.Close()
-		data, _ := ioutil.ReadAll(res.Body)
+	defer res.Body.Close()
+	body, _ := ioutil.ReadAll(res.Body)
 
-		expectedResponse := fmt.Sprintf("[{\"name\":\"ShowRSS\",\"url\":\"%v\"}]\n", rssAddress)
-		assert.Equal(expectedResponse, string(data))
-	})
-
-	t.Run("Episodes endpoint", func(t *testing.T) {
-		episodes := []trss.Episode{
-			mocks.EpisodeExample,
-		}
-
-		mockEpisodes.MockAll = func() ([]trss.Episode, error) {
-			return episodes, nil
-		}
-
-		req := httptest.NewRequest(http.MethodGet, "/", nil)
-		w := httptest.NewRecorder()
-		api := new(trss.Api)
-
-		api.Episodes(w, req)
-
-		res := w.Result()
-
-		defer res.Body.Close()
-		data, _ := ioutil.ReadAll(res.Body)
-
-		var resultData []trss.Episode
-		jsonErr := json.Unmarshal(data, &resultData)
-
-		if jsonErr != nil {
-			t.Error(jsonErr)
-		}
-
-		assert.Equal(episodes, resultData)
-	})
-
-	t.Run("Episodes endpoint fail", func(t *testing.T) {
-		episodes := []trss.Episode{}
-
-		mockEpisodes.MockAll = func() ([]trss.Episode, error) {
-			return episodes, errors.New("Database error")
-		}
-
-		req := httptest.NewRequest(http.MethodGet, "/", nil)
-		w := httptest.NewRecorder()
-		api := new(trss.Api)
-
-		api.Episodes(w, req)
-
-		res := w.Result()
-
-		assert.Equal(500, res.StatusCode)
-
-		defer res.Body.Close()
-		data, _ := ioutil.ReadAll(res.Body)
-
-		var resultData map[string]string
-		jsonErr := json.Unmarshal(data, &resultData)
-
-		if jsonErr != nil {
-			t.Error(jsonErr)
-		}
-
-		expectedResponse := map[string]string{"error": "Could not read episodes"}
-
-		assert.Equal(expectedResponse, resultData)
-	})
-
-	t.Run("Download endpoint", func(t *testing.T) {
-		mockFeeds.MockFetchItems = func(r string) ([]trss.FeedItem, error) {
-			feedItems := []trss.FeedItem{mocks.FeedItemExample}
-
-			return feedItems, nil
-		}
-
-		mockEpisodes.MockFindEpisode = func(e *trss.Episode) (trss.Episode, error) {
-			return trss.Episode{}, gorm.ErrRecordNotFound
-		}
-
-		mockEpisodes.MockDownloadEpisode = func(e trss.Episode) error {
-			return nil
-		}
-
-		mockEpisodes.MockAddEpisode = func(e *trss.Episode) error {
-			e.ID = 2
-			return nil
-		}
-
-		req := httptest.NewRequest(http.MethodGet, "/", nil)
-		w := httptest.NewRecorder()
-		api := new(trss.Api)
-
-		api.Download(w, req)
-
-		res := w.Result()
-
-		assert.Equal(200, res.StatusCode)
-
-		defer res.Body.Close()
-		data, _ := ioutil.ReadAll(res.Body)
-
-		type ResponseType struct {
-			Episodes []trss.Episode
-			Errors   []string
-		}
-
-		expectedEpisode := mocks.EpisodeExample
-		expectedEpisode.ID = 2
-		expectedEpisodes := []trss.Episode{expectedEpisode}
-		expectedErrors := []string(nil)
-
-		responseData := &ResponseType{}
-
-		json.Unmarshal(data, responseData)
-
-		assert.Equal(expectedEpisodes, responseData.Episodes)
-		assert.Equal(expectedErrors, responseData.Errors)
-	})
-
-	t.Run("Download endpoint fail", func(t *testing.T) {
-		mockFeeds.MockFetchItems = func(r string) ([]trss.FeedItem, error) {
-			feedItems := []trss.FeedItem{mocks.FeedItemExample}
-
-			return feedItems, nil
-		}
-
-		mockEpisodes.MockFindEpisode = func(e *trss.Episode) (trss.Episode, error) {
-			return trss.Episode{}, gorm.ErrRecordNotFound
-		}
-
-		mockEpisodes.MockDownloadEpisode = func(e trss.Episode) error {
-			return errors.New("Transmission error")
-		}
-
-		mockEpisodes.MockAddEpisode = func(e *trss.Episode) error {
-			e.ID = 2
-			return nil
-		}
-
-		req := httptest.NewRequest(http.MethodGet, "/", nil)
-		w := httptest.NewRecorder()
-		api := new(trss.Api)
-
-		api.Download(w, req)
-
-		res := w.Result()
-
-		assert.Equal(200, res.StatusCode)
-
-		defer res.Body.Close()
-		data, _ := ioutil.ReadAll(res.Body)
-
-		type ResponseType struct {
-			Episodes []trss.Episode
-			Errors   []string
-		}
-
-		expectedEpisodes := []trss.Episode(nil)
-		expectedErrors := []string{"Transmission error"}
-
-		responseData := &ResponseType{}
-
-		json.Unmarshal(data, responseData)
-
-		assert.Equal(expectedEpisodes, responseData.Episodes)
-		assert.Equal(expectedErrors, responseData.Errors)
-	})
-
-	t.Run("Clean endpoint", func(t *testing.T) {
-		mockTrs.MockCleanFinished = func() ([]string, error) {
-			return []string{"Episode title"}, nil
-		}
-
-		req := httptest.NewRequest(http.MethodGet, "/", nil)
-		w := httptest.NewRecorder()
-		api := new(trss.Api)
-
-		api.Clean(w, req)
-
-		res := w.Result()
-
-		assert.Equal(200, res.StatusCode)
-
-		defer res.Body.Close()
-		data, _ := ioutil.ReadAll(res.Body)
-
-		responseData := []string{}
-
-		json.Unmarshal(data, &responseData)
-
-		expectedResponse := []string{"Episode title"}
-
-		assert.Equal(expectedResponse, responseData)
-	})
-
-	t.Run("Clean endpoint fail", func(t *testing.T) {
-		mockTrs.MockCleanFinished = func() ([]string, error) {
-			return []string{}, errors.New("Remove error")
-		}
-
-		req := httptest.NewRequest(http.MethodGet, "/", nil)
-		w := httptest.NewRecorder()
-		api := new(trss.Api)
-
-		api.Clean(w, req)
-
-		res := w.Result()
-
-		assert.Equal(500, res.StatusCode)
-
-		defer res.Body.Close()
-		data, _ := ioutil.ReadAll(res.Body)
-
-		responseData := map[string]string{}
-
-		json.Unmarshal(data, &responseData)
-
-		expectedResponse := map[string]string{"error": "Could not remove torrents"}
-
-		assert.Equal(expectedResponse, responseData)
-	})
+	return res, body
 }
+
+var _ = Describe("Api", func() {
+	var api *trss.Api
+	var rssAddress string
+	var mockFeeds *mocks.MockFeeds
+	var mockEpisodes *mocks.MockEpisodes
+	var mockTrs *mocks.MockTransmissionService
+
+	BeforeEach(func() {
+		api = new(trss.Api)
+		rssAddress = "https://example.rss/feed"
+		os.Setenv("RSS_FEED_ADDRESS", rssAddress)
+		DeferCleanup(func() {
+			os.Unsetenv("RSS_FEED_ADDRESS")
+		})
+		mockFeeds = &mocks.MockFeeds{}
+		trss.FeedService = mockFeeds
+		mockEpisodes = &mocks.MockEpisodes{}
+		trss.EpisodeService = mockEpisodes
+		mockTrs = &mocks.MockTransmissionService{}
+		trss.TrsService = mockTrs
+	})
+
+	Context("Feeds endpoint", func() {
+		It("should return feeds", func() {
+			_, body := sendApiRequest(api.Feeds, http.MethodGet, "/")
+
+			var result []interface{}
+			err := json.Unmarshal(body, &result)
+			Expect(err).To(BeNil())
+
+			expectedResult := []interface{}{map[string]interface{}{"name": "ShowRSS", "url": rssAddress}}
+
+			Expect(result).To(Equal(expectedResult))
+		})
+	})
+
+	Context("Episodes endpoint", func() {
+		It("should return episodes", func() {
+			episodes := []trss.Episode{
+				mocks.EpisodeExample,
+			}
+
+			mockEpisodes.MockAll = func() ([]trss.Episode, error) {
+				return episodes, nil
+			}
+
+			_, body := sendApiRequest(api.Episodes, http.MethodGet, "/")
+
+			var resultData []trss.Episode
+			jsonErr := json.Unmarshal(body, &resultData)
+
+			Expect(jsonErr).To(BeNil())
+
+			Expect(resultData).To(Equal(episodes))
+		})
+
+		It("shoud returns HTTP 500 if query fails", func() {
+			episodes := []trss.Episode{}
+
+			mockEpisodes.MockAll = func() ([]trss.Episode, error) {
+				return episodes, errors.New("Database error")
+			}
+
+			res, body := sendApiRequest(api.Episodes, http.MethodGet, "/")
+
+			Expect(res.StatusCode).To(Equal(500))
+
+			var resultData map[string]string
+			jsonErr := json.Unmarshal(body, &resultData)
+
+			Expect(jsonErr).To(BeNil())
+			Expect(resultData).To(HaveKeyWithValue("error", "Could not read episodes"))
+		})
+	})
+
+	Context("Download endpoint", func() {
+		It("should respond whith downloaded items", func() {
+
+		})
+
+		It("should resturn error", func() {
+
+		})
+	})
+})
